Move a couple functions into their own files to prep for removing WorkflowHost.
