+++ kotlin/workflow-runtime/src/main/java/com/squareup/workflow/WorkflowHost.kt
@@ -16,21 +16,13 @@
 package com.squareup.workflow
 
 import com.squareup.workflow.WorkflowHost.Factory
-import com.squareup.workflow.WorkflowHost.RenderingAndSnapshot
-import com.squareup.workflow.internal.WorkflowNode
-import com.squareup.workflow.internal.id
+import com.squareup.workflow.internal.runWorkflowLoop
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.channels.consume
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.produceIn
-import kotlinx.coroutines.selects.select
 import org.jetbrains.annotations.TestOnly
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
@@ -47,14 +39,6 @@ private val DEFAULT_WORKFLOW_COROUTINE_NAME = CoroutineName("WorkflowHost")
  * it.
  */
 interface WorkflowHost<out OutputT : Any, out RenderingT> {
-
-  /**
-   * Emitted from [renderingsAndSnapshots] after every render pass.
-   */
-  data class RenderingAndSnapshot<out RenderingT>(
-    val rendering: RenderingT,
-    val snapshot: Snapshot
-  )
 
   /**
    * Stream of [renderings and snapshots][RenderingAndSnapshot] from the root workflow.
@@ -135,7 +119,7 @@ interface WorkflowHost<out OutputT : Any, out RenderingT> {
         // Put the coroutine name first so the passed-in contexts can override it.
         context = DEFAULT_WORKFLOW_COROUTINE_NAME + baseContext + context
     ) { onRendering, onOutput ->
-      runWorkflowTree(
+      runWorkflowLoop(
           workflow = workflow.asStatefulWorkflow(),
           inputs = inputs,
           initialSnapshot = snapshot,
@@ -168,7 +152,7 @@ interface WorkflowHost<out OutputT : Any, out RenderingT> {
     ): WorkflowHost<OutputT, RenderingT> = RealWorkflowHost(
         context = DEFAULT_WORKFLOW_COROUTINE_NAME + baseContext
     ) { onRendering, onOutput ->
-      runWorkflowTree(
+      runWorkflowLoop(
           workflow = workflow.asStatefulWorkflow(),
           inputs = inputs,
           initialSnapshot = null,
@@ -176,82 +160,6 @@ interface WorkflowHost<out OutputT : Any, out RenderingT> {
           onRendering = onRendering,
           onOutput = onOutput
       )
-    }
-  }
-}
-
-/**
- * Loops forever, or until the coroutine is cancelled, processing the workflow tree and emitting
- * updates by calling [onRendering] and [onOutput].
- *
- * This function is the lowest-level entry point into the runtime. Don't call this directly, instead
- * use [WorkflowHost.Factory] to create a [WorkflowHost].
- */
-@UseExperimental(FlowPreview::class, ExperimentalCoroutinesApi::class)
-suspend fun <InputT, StateT, OutputT : Any, RenderingT> runWorkflowTree(
-  workflow: StatefulWorkflow<InputT, StateT, OutputT, RenderingT>,
-  inputs: Flow<InputT>,
-  initialSnapshot: Snapshot?,
-  initialState: StateT? = null,
-  onRendering: suspend (RenderingAndSnapshot<RenderingT>) -> Unit,
-  onOutput: suspend (OutputT) -> Unit
-): Nothing = coroutineScope {
-  val inputsChannel = inputs.produceIn(this)
-  inputsChannel.consume {
-    var output: OutputT? = null
-    var input: InputT = inputsChannel.receive()
-    var inputsClosed = false
-    val rootNode = WorkflowNode(
-        id = workflow.id(),
-        workflow = workflow,
-        initialInput = input,
-        snapshot = initialSnapshot,
-        baseContext = coroutineContext,
-        initialState = initialState
-    )
-
-    try {
-      while (true) {
-        coroutineContext.ensureActive()
-
-        val rendering = rootNode.render(workflow, input)
-        val snapshot = rootNode.snapshot(workflow)
-
-        onRendering(RenderingAndSnapshot(rendering, snapshot))
-        output?.let { onOutput(it) }
-
-        // Tick _might_ return an output, but if it returns null, it means the state or a child
-        // probably changed, so we should re-render/snapshot and emit again.
-        output = select {
-          // Stop trying to read from the inputs channel after it's closed.
-          if (!inputsClosed) {
-            @Suppress("EXPERIMENTAL_API_USAGE")
-            inputsChannel.onReceiveOrNull { newInput ->
-              if (newInput == null) {
-                inputsClosed = true
-              } else {
-                input = newInput
-              }
-              // No output. Returning from the select will go to the top of the loop to do another
-              // render pass.
-              return@onReceiveOrNull null
-            }
-          }
-
-          // Tick the workflow tree.
-          rootNode.tick(this) { it }
-        }
-      }
-      // Compiler gets confused, and thinks both that this throw is unreachable, and without the
-      // throw that the infinite while loop will exit normally and thus need a return statement.
-      @Suppress("UNREACHABLE_CODE", "ThrowableNotThrown")
-      throw AssertionError()
-    } finally {
-      // There's a potential race condition if the producer coroutine is cancelled before it has a
-      // chance to enter the try block, since we can't use CoroutineStart.ATOMIC. However, until we
-      // actually see this cause problems, I'm not too worried about it.
-      // See https://github.com/Kotlin/kotlinx.coroutines/issues/845
-      rootNode.cancel()
     }
   }
 }
